+++ app/src/main/java/woowacourse/shopping/data/datasource/remote/RecentProductLocalDataSource.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.datasource
+package woowacourse.shopping.data.datasource.remote
 
 import woowacourse.shopping.data.db.RecentProductEntity
