+++ app/src/main/java/woowacourse/shopping/di/provider/DataSourceProvider.kt
@@ -1,14 +1,14 @@
 package woowacourse.shopping.di.provider
 
 import woowacourse.shopping.ShoppingApplication
-import woowacourse.shopping.data.datasource.CartLocalDataSource
-import woowacourse.shopping.data.datasource.CartLocalDataSourceImpl
-import woowacourse.shopping.data.datasource.CartRemoteDataSource
-import woowacourse.shopping.data.datasource.CartRemoteDataSourceImpl
-import woowacourse.shopping.data.datasource.ProductRemoteDataSource
-import woowacourse.shopping.data.datasource.ProductRemoteDataSourceImpl
-import woowacourse.shopping.data.datasource.RecentProductLocalDataSource
-import woowacourse.shopping.data.datasource.RecentProductLocalDataSourceImpl
+import woowacourse.shopping.data.datasource.local.CartLocalDataSource
+import woowacourse.shopping.data.datasource.local.CartLocalDataSourceImpl
+import woowacourse.shopping.data.datasource.remote.CartRemoteDataSource
+import woowacourse.shopping.data.datasource.remote.CartRemoteDataSourceImpl
+import woowacourse.shopping.data.datasource.remote.ProductRemoteDataSource
+import woowacourse.shopping.data.datasource.remote.ProductRemoteDataSourceImpl
+import woowacourse.shopping.data.datasource.remote.RecentProductLocalDataSource
+import woowacourse.shopping.data.datasource.remote.RecentProductLocalDataSourceImpl
 import woowacourse.shopping.data.db.ShoppingDatabase
 
 object DataSourceProvider {
